change type of var

+++ Server/src/main/java/com/example/cryptoworld/service/CryptoService.java
@@ -13,5 +13,5 @@ public interface CryptoService {
 
     BigDecimal cryptoMarketCap4Hour ();
 
-    double calculateTheMarketDifferenceInPercentage();
+    BigDecimal calculateTheMarketDifferenceInPercentage();
 }
